chore: use .env over next.js's preferred .env.local
This is because there are values in .env that are shared between
our app and docker compose. Docker compose only uses the current
environment and .env (not .env.local) for variable substitutions
in the compose file. There are workarounds using either the
--env-file docker compose CLI option or env_file option in the
compose file, but these are not ideal:

* --env-file cannot be specified for devcontainers

* env_file loads everything from the specified env file into the
  container, rather than allowing us to specify variable
  substitutions in the compose file like we can with .env

The approach with the least hassle is to just use .env. Next.js loads
that anyway. The example file's called .env.example instead.